apps/dashboard-backend/pyproject.toml
[project]
name = "dashboard-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "snip-db[pg]",
    "fastapi[standard]",
    "pydantic-settings",
    "httpx",
    "python-jose[cryptography]",
    "shortuuid",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "pytest-cov", "httpx", "ruff", "aiosqlite"]

[tool.uv.sources]
snip-db = { workspace = true }

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B008"]  # B008: Depends() in function defaults is standard FastAPI pattern

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
